+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - updated to 26.1.x
 
+## Fixes
+
+- fixed an issue where the "natural player attribute values" where applied to every LivingEntity
+
 # 1.0.0
 
 First release.